refactor: consolidated error message parsing
