+++ src/hooks/lending/useRepayOperations.ts
@@ -6,7 +6,7 @@ import { ethers } from "ethers";
 import { evmAddress, bigDecimal } from "@aave/react";
 import { useAaveRepay, useAavePermit } from "@/hooks/aave/useAaveInteractions";
 import { useChainSwitch } from "@/utils/swap/walletMethods";
-import { truncateAddress } from "@/utils/formatters";
+import { truncateAddress, parseDepositError } from "@/utils/formatters";
 import { UnifiedReserveData, ChainId } from "@/types/aave";
 import { Chain, Token } from "@/types/web3";
 import { getChainByChainId } from "@/config/chains";
@@ -171,16 +171,15 @@ export const useRepayOperations = (
           console.error("Repay failed:", result.error);
           toast.error("Repay failed", {
             id: repayToastId,
-            description: result.error || "An unknown error occurred",
+            description: parseDepositError(
+              result.error || "An unknown error occurred",
+            ),
           });
         }
       } catch (error) {
         console.error("Repay operation failed:", error);
         toast.error("Repay operation failed", {
-          description:
-            error instanceof Error
-              ? error.message
-              : "An unknown error occurred",
+          description: parseDepositError(error),
         });
       }
     },
